Remove underscore character from DataMigrator’s name
